관심사의 분리 : RestaurantInputContainer에 대한 테스트를 추가한다
